+++ src/pages/instructorRegister/ImageUploader/index.tsx
@@ -1,50 +1,10 @@
-import { useEffect, useRef, useState } from 'react';
 import Flex from '@/components/Flex';
+import { useUploadImg } from '@/hooks/useUploadImg';
 import { IcPlusGray0524, IcXCircleGray0424 } from '@/assets/svg';
 import { closeIconStyle, inputStyle, previewImgStyle } from './index.css';
 
 const ImageUploader = () => {
-  const [previewImg, setPreviewImg] = useState('');
-  const [imgFile, setImgFile] = useState<File>();
-  const imgRef = useRef<HTMLInputElement | null>(null);
-
-  const handleUploaderClick = () => {
-    if (imgRef.current) {
-      imgRef.current.click();
-    }
-  };
-
-  const uploadImgFile = () => {
-    if (!imgRef.current || !imgRef.current.files) {
-      return;
-    }
-    const file = imgRef.current.files[0];
-    const reader = new FileReader();
-
-    setImgFile(file);
-
-    reader.readAsDataURL(file);
-    reader.onloadend = () => {
-      if (typeof reader.result === 'string') {
-        setPreviewImg(reader.result);
-      }
-    };
-  };
-
-  const deleteImgFile = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    setImgFile(undefined);
-    setPreviewImg('');
-
-    if (imgRef.current) {
-      imgRef.current.value = '';
-    }
-  };
-
-  useEffect(() => {
-    console.log(imgFile);
-  }, [imgFile]);
-
+  const { imgRef, previewImg, imgFile, handleUploaderClick, uploadImgFile, deleteImgFile } = useUploadImg();
   return (
     <>
       <Flex
